server: Extract action-server.js into its own module
Change-type: patch
