+++ lib/environment/index.js
@@ -91,10 +91,8 @@ exports.getDatabaseConfiguration = () => {
 		password: process.env.DB_PASSWORD || '',
 		port: process.env.DB_PORT,
 		host: process.env.DB_HOST,
-		name: process.env.SERVER_DATABASE || 'jellyfish',
-		pool: {
-			minimumSize: process.env.RETHINKDB_MIN_POOL_SIZE,
-			maximumSize: process.env.RETHINKDB_MAX_POOL_SIZE
-		}
+		database: process.env.SERVER_DATABASE || 'jellyfish',
+		buffer: process.env.RETHINKDB_MIN_POOL_SIZE,
+		max: process.env.RETHINKDB_MAX_POOL_SIZE
 	}
 }
