+++ lib/server/bootstrap.js
@@ -22,8 +22,8 @@ const actionLibrary = require('../action-library')
 const errorReporter = require('../error-reporter')
 const environment = require('../environment')
 const logger = require('../logger').getLogger(__filename)
+const actionServer = require('../action-server')
 
-const actionServer = require('./action-server')
 const configuration = require('./configuration')
 const cardLoader = require('./card-loader')
 const http = require('./http')
