+++ lib/server/http/routes.js
@@ -22,7 +22,7 @@ const uuid = require('uuid/v4')
 const errio = require('errio')
 const multer = require('multer')
 const randomstring = require('randomstring')
-const errorReporter = require('../error-reporter')
+const errorReporter = require('../../error-reporter')
 const Storage = require('./file-storage')
 const logger = require('../../logger').getLogger(__filename)
 const environment = require('../../environment')
